Working Code for SpaceFlow
Contains working code for SpaceFlow clustering method using the DLPFC 1 dataset. To change the dataset just change the name in the 3rd code block. To reproduce the result set up an environment with python3.11.4 and then follow the tutorial on the SpaceFlow github. If any networkx errors occur change the version to 2.8 using:
pip install networkx==2.8